+++ tests/Integration/Controller/ManufacturerWithTokenTest.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace OxidEsales\GraphQL\Catalogue\Tests\Integration\Controller;
 
 use DateTimeImmutable;
-use OxidEsales\GraphQL\Catalogue\Tests\Integration\TokenTestCase;
+use OxidEsales\GraphQL\Base\Tests\Integration\TokenTestCase;
 use TheCodingMachine\GraphQLite\Types\DateTimeType;
 
 final class ManufacturerWithTokenTest extends TokenTestCase