+++ tests/Integration/DataType/ProductRelationServiceTest.php
@@ -12,7 +12,7 @@ namespace OxidEsales\GraphQL\Catalogue\Tests\Integration\DataType;
 use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
 use OxidEsales\EshopCommunity\Internal\Framework\Database\QueryBuilderFactoryInterface;
-use OxidEsales\GraphQL\Catalogue\Tests\Integration\TokenTestCase;
+use OxidEsales\GraphQL\Base\Tests\Integration\TokenTestCase;
 
 /**
  * @covers OxidEsales\GraphQL\Catalogue\DataType\ProductRelationService