+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Repository::delete() takes loaded BaseModel as argument.
 
+### Removed
+
+- `\OxidEsales\GraphQL\Catalogue\Tests\Integration\TokenTestCase`
+
 ## [0.1.0] 2020-06-23
 
 Initial release 🎉